+++ inbas-reporting-tool/src/main/java/cz/cvut/kbss/inbas/reporting/persistence/dao/OccurrenceReportDao.java
@@ -4,12 +4,12 @@ import cz.cvut.kbss.inbas.reporting.model.Occurrence;
 import cz.cvut.kbss.inbas.reporting.model.OccurrenceReport;
 import cz.cvut.kbss.inbas.reporting.model.Vocabulary;
 import cz.cvut.kbss.inbas.reporting.persistence.dao.util.OrphanRemover;
+import cz.cvut.kbss.jopa.exceptions.NoResultException;
 import cz.cvut.kbss.jopa.model.EntityManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import java.net.URI;
-import java.util.List;
 
 @Repository
 public class OccurrenceReportDao extends BaseReportDao<OccurrenceReport> implements GenericDao<OccurrenceReport> {
@@ -67,29 +67,19 @@ public class OccurrenceReportDao extends BaseReportDao<OccurrenceReport> impleme
      * @param occurrence The occurrence to filter reports by
      * @return List of reports
      */
-    public List<OccurrenceReport> findByOccurrence(Occurrence occurrence) {
+    public OccurrenceReport findByOccurrence(Occurrence occurrence) {
         final EntityManager em = entityManager();
         try {
             return em.createNativeQuery(
                     "SELECT ?x WHERE { ?x a ?type ;" +
-                            "?hasRevision ?revision ;" +
-                            "?hasFileNumber ?fileNo ;" +
-                            "?documents ?occurrence . " +
-                            "?occurrence ?hasStartTime ?startTime ." +
-                            // Use only the max revision reports
-                            "{ SELECT (MAX(?rev) AS ?maxRev) ?iFileNo WHERE " +
-                            "{ ?y a ?type ; ?documents ?occurrence ; ?hasFileNumber ?iFileNo ; ?hasRevision ?rev . }" +
-                            " GROUP BY ?iFileNo }" +
-                            "FILTER (?revision = ?maxRev && ?fileNo = ?iFileNo)" +
-                            "} ORDER BY DESC(?startTime) DESC(?revision)",
+                            "?documents ?occurrence . }",
                     OccurrenceReport.class)
                      .setParameter("type", typeIri)
-                     .setParameter("occurrence", occurrence.getUri())
-                     .setParameter("hasRevision", URI.create(Vocabulary.s_p_has_revision))
                      .setParameter("documents", URI.create(Vocabulary.s_p_documents))
-                     .setParameter("hasStartTime", URI.create(Vocabulary.s_p_has_start_time))
-                     .setParameter("hasFileNumber", URI.create(Vocabulary.s_p_has_file_number))
-                     .getResultList();
+                     .setParameter("occurrence", occurrence.getUri())
+                     .getSingleResult();
+        } catch (NoResultException e) {
+            return null;
         } finally {
             em.close();
         }